Rename tickets/internal package
